[MOD] moved creation of popup in account pages, communication via props with popups and buttons, modification of css
