+++ src/Components/Accounts/SchoolAdm/SchoolAccountsTable.js
@@ -2,50 +2,25 @@ import '../../../css/Components/Accounts/AccountsTable.css'
 import { React, useState, useEffect } from 'react'
 
 export default function SchoolAccountsTable () {
-  const [teacherAccountList, setTeacherAccountList] = useState([]) // list of accounts
-  const [studentAccountList, setStudentAccountList] = useState([]) // list of accounts
+  const [accountList, setAccountList] = useState([]); // list of accounts
 
 
   // get request for account list
   async function getAccountList () {
-    const baseUrl = process.env.REACT_APP_BACKEND_URL + '/api/profiles/by/'
+    const baseUrl = process.env.REACT_APP_BACKEND_URL + '/user/all'
     const token = sessionStorage.getItem('token')
 
-    const respTeacher = await fetch(baseUrl + 'teacher', {
+    const resp = await fetch(baseUrl, {
       method: 'GET',
       headers: {
         'x-auth-token': token
       }
     })
-    console.log(respTeacher);
-    const tdata = await respTeacher.json()
+    console.log(resp);
+    const data = await resp.json();
+    console.log(data);
 
-    const respStudent = await fetch(baseUrl + 'student', {
-      method: 'GET',
-      headers: {
-        'x-auth-token': token
-      }
-    })
-    console.log(respStudent);
-    const sdata = await respStudent.json()
-    // const data = [
-    //   {
-    //     'firstName': 'Teacher',
-    //     'lastName': '1',
-    //     'email': 'teacher1@schood.fr',
-    //     'role': 'Teacher',
-    //     'classe': '1'
-    //   },
-    //   {
-    //     'firstName': 'Student',
-    //     'lastName': '1',
-    //     'email': 'student1@schood.fr',
-    //     'role': 'Student',
-    //     'classe': '1'
-    //   },
-    // ];
-    setTeacherAccountList(tdata);
-    setStudentAccountList(sdata);
+    setAccountList(data);
   }
 
   // account list request on mounted
@@ -68,17 +43,17 @@ export default function SchoolAccountsTable () {
           </thead>
           <tbody id='tableBody'>
             {
-              teacherAccountList.map((data, index) =>
+              accountList.map((data, index) =>
                 <tr key={index}>
-                  <td>{data.firstName}</td>
-                  <td>{data.lastName}</td>
+                  <td>{data.firstname}</td>
+                  <td>{data.lastname}</td>
                   <td>{data.email}</td>
-                  <td>{data.role}</td>
-                  <td>{data.classe}</td>
+                  <td>{data.role.name}</td>
+                  <td>{data.classes.name}</td>
                 </tr>
               )
             }
-            {
+            {/* {
               studentAccountList.map((data, index) =>
                 <tr key={index}>
                   <td>{data.firstName}</td>
@@ -88,7 +63,7 @@ export default function SchoolAccountsTable () {
                   <td>{data.classe}</td>
                 </tr>
               )
-            }
+            } */}
           </tbody>
         </table>
       </div>
